Wrap usage of SimulationParametersRepository to SimulationParametersScreenController

+++ Structural_Balance-Symulations_and_Visualisation/src/main/java/com/example/project/database/model/SimulationParameters.java
@@ -11,12 +11,12 @@ public class SimulationParameters {
     @Column(name = "number_of_steps")
     private int numberOfSteps;
     @Column(name = "annealing_Parameter")
-    private int annealingParameter;
+    private double annealingParameter;
 
     public SimulationParameters() {
     }
 
-    public SimulationParameters(int numberOfSteps, int annealingParameter) {
+    public SimulationParameters(int numberOfSteps, double annealingParameter) {
         this.numberOfSteps = numberOfSteps;
         this.annealingParameter = annealingParameter;
     }
@@ -37,11 +37,11 @@ public class SimulationParameters {
         this.numberOfSteps = numberOfSteps;
     }
 
-    public int getAnnealingParameter() {
+    public double getAnnealingParameter() {
         return annealingParameter;
     }
 
-    public void setAnnealingParameter(int annealingParameter) {
+    public void setAnnealingParameter(double annealingParameter) {
         this.annealingParameter = annealingParameter;
     }
 }
